+++ src/main/java/com/netlify/hparcells/chaos/events/FullHealEvent.java
@@ -1,0 +1,20 @@
+package com.netlify.hparcells.chaos.events;
+
+import com.netlify.hparcells.chaos.Chaos;
+import com.netlify.hparcells.chaos.ChaosEvent;
+import org.bukkit.Bukkit;
+import org.bukkit.entity.Player;
+
+public class FullHealEvent extends ChaosEvent {
+    public void onEnable(String username, Chaos plugin) {
+        Player target = (Bukkit.getServer().getPlayer(username));
+
+        if(target != null) {
+            target.setHealth(20);
+        }
+    }
+
+    public String getName() {
+        return "Full Heal";
+    }
+}