+++ src/main/java/com/netlify/hparcells/chaos/events/ZeroHungerEvent.java
@@ -1,0 +1,21 @@
+package com.netlify.hparcells.chaos.events;
+
+import com.netlify.hparcells.chaos.Chaos;
+import com.netlify.hparcells.chaos.ChaosEvent;
+import org.bukkit.Bukkit;
+import org.bukkit.entity.Player;
+
+public class ZeroHungerEvent extends ChaosEvent {
+    public void onEnable(String username, Chaos plugin) {
+        Player target = (Bukkit.getServer().getPlayer(username));
+
+        if(target != null) {
+            target.setFoodLevel(0);
+            target.setSaturation(0f);
+        }
+    }
+
+    public String getName() {
+        return "Zero Hunger";
+    }
+}